bugfix(Calculator,CalculatorUtil) 숫자 하나를 문자열로 입력할 경우 해당 숫자를 반환한다
기존 로직은 10을 입력하면 필터링 되지 못했음,
MagicNumber 0 -> ZeroValue로 치환

+++ src/test/java/calculator/CalculatorUtil.java
@@ -12,7 +12,7 @@ public class CalculatorUtil {
 
     public static boolean isNumeric(String text) {
         try {
-            Double.parseDouble(text);
+            Integer.parseInt(text);
         } catch (NumberFormatException e) {
             return false;
         }
